Trait p25::bits::IterParams

pub trait IterParams {
    type Input;
    type Output;
    fn bits() -> usize;

    fn buffer() -> usize;

    fn shift() -> usize;

    fn to_byte(input: Self::Input) -> u8;

    fn to_output(bits: u8) -> Self::Output;

    fn post_shift() -> usize { ... }

    fn iterations() -> usize { ... }

    fn validate() { ... }
}

Associated Types

type Input

Type to consume when buffering.

type Output

Type to yield at each iteration.

Required Methods

fn bits() -> usize

Number of bits to consume at each iteration.

fn buffer() -> usize

Number of input symbols to consume when buffering.

fn shift() -> usize

Amount to shift buffer after loading an input symbol.

fn to_byte(input: Self::Input) -> u8

Convert input symbol to a byte.

fn to_output(bits: u8) -> Self::Output

Convert bits to output type.

Provided Methods

fn post_shift() -> usize

Amount to shift buffer after all buffering, so the bits are lined up at the MSB.

fn iterations() -> usize

Number of iterations before buffering.

fn validate()

Verify the parameters are supported.