Struct p25::coding::bmcf::ErrorDescriptions

pub struct ErrorDescriptions<P: PolynomialCoefs> { /* fields omitted */ }

Computes error locations and patterns from the roots of the error locator polynomial Λ(x).

This uses the Forney algorithm for error pattern evaluation, which avoids polynomial long division.

Methods

impl<P: PolynomialCoefs> ErrorDescriptions<P>

fn new(syn: Polynomial<P>, loc: Polynomial<P>) -> Self

Create a new ErrorDescriptions from the given syndrome polynomial s(x) and error locator polynomial Λ(x).

fn for_root(&self, root: P25Codeword) -> (usize, P25Codeword)

Compute the error location and pattern for the given root ai-1 of Λ(x).