Struct p25::coding::bmcf::ErrorLocs

pub struct ErrorLocs<P: PolynomialCoefs> { /* fields omitted */ }

Uses Berlekamp-Massey algorithm to find the error location polynomial Λ(x) from the syndrome polynomial s(x).

The error location polynomial is defined as Λ(x) = (1 + xa1)(1 + xa2) ··· (1 + xa2t) for error locations a1, ..., ae, where e ≤ t is the number of detected errors. Each ai = αki for some ki. The roots of Λ(x) are then a1-1, ..., ae-1, i.e., if Λ(αk) = 0, then an error occured at symbol m in the received word, where α-k ≡ αm in the field.

where t = floor((d - 1) / 2) and d is from the defining (n, k, d).

s(x) = 1 + s1x + s2x2 + ··· + s2tx2t, where si = r(αi) using the received codeword polynomial r(x).

Methods

impl<P: PolynomialCoefs> ErrorLocs<P>

fn new(syn: Polynomial<P>) -> ErrorLocs<P>

Construct a new ErrorLocs from the given syndrome polynomial.

fn decode(self) -> Polynomial<P>

Perform the iterative steps to get the error-location polynomial Λ(x) wih deg(Λ) <= P::errors().