Struct p25::coding::galois::Codeword

pub struct Codeword<F: GaloisField> { /* fields omitted */ }

Codeword in a Galois field.

Methods

impl<F: GaloisField> Codeword<F>

fn new(bits: u8) -> Codeword<F>

Construct a new Codeword αi from the given bit pattern. Panic if the pattern is invalid in the field.

fn for_power(power: usize) -> Codeword<F>

Construct a new Codeword αm ≡ αi (modulo the field) for the given power i.

fn bits(&self) -> u8

Retrieve the bit pattern of the codeword.

fn zero(&self) -> bool

Check if the codeword is zero.

fn power(&self) -> Option<usize>

Retrieve the power i of the current codeword αi. Return Some(i) if the power is defined and None if the codeword is zero.

fn invert(self) -> Codeword<F>

Find 1/αi for the current codeword αi. Panic if the codeword is zero.

fn pow(&self, pow: usize) -> Codeword<F>

Compute (αi)p for the current codeword αi and given power p.