Module p25::consts

Constants used throughout the protocol.

Constants

CODING_DIBITS
Number of dibits that are input to the 1/2 or 3/4-rate trellis coder.

CRYPTO_CONTROL_BYTES
Number of bytes in a crypto control word.

DATA_FRAG_DIBITS
Number of dibits in the voice data fragment.

EXTRA_HEXBITS
Number of hexbits in a coded voice extra packet.

EXTRA_PIECE_DIBITS
Number of dibits in an LC/CC piece. An LC/CC word is spread over 6 equal-sized pieces in each frame group, for a total of 120 dibits.

EXTRA_WORD_DIBITS
Number of dibits in each coded word that makes up a voice extra component.

FRAME_DIBITS
Number of dibits in a coded voice frame.

HEADER_BYTES
Number of bytes in an uncoded voice header packet.

HEADER_HEXBITS
Number of hexbits in a coded voice header packet.

HEADER_WORD_DIBITS
Number of dibits in each coded word that makes up the voice header packet.

LC_TERM_WORD_DIBITS
Number of dibits in each coded word that makes up the voice LC terminator packet.

LINK_CONTROL_BYTES
Number of bytes in a link control word.

NID_DIBITS
Number of dibits in a coded NID word.

SAMPLE_RATE
Baseband samples per second

SYMBOL_PERIOD
Baseband samples per symbol.

SYMBOL_RATE
Symbols (dibits) per second.

SYNC_SYMBOLS
Number of symbols in the frame sync sequence.

TSBK_BYTES
Number of bytes in an uncoded TSBK packet.

TSBK_DIBITS
Number of dibits in an uncoded TSBK packet.