Module p25::data::crc

Implements the 9, 16, and 32-bit CRCs defined by P25 for data checksums.

This implementation uses the typical long division and takes advantage of the short lengths to use only a 64-bit word as a buffer, allowing simple bitwise operations.

Structs

CRC
CRC calculator using long division.

CRC16Params
Params for 16-bit CRC.

CRC32Params
Params for 32-bit CRC.

CRC9Params
Params for 9-bit CRC.

Traits

CRCParams

Type Definitions

CRC16
16-bit CRC calculator.

CRC32
32-bit CRC calculator.

CRC9
9-bit CRC calculator.