Module p25::data::fragment

Provides Fragments iterator for splitting a data slice into fragment slices, whose size depends on the data packet used.

A data message is made up of unlimited data fragments, with the FSNF field of the packet header determining how the message is reconstructed from the fragments.

Structs

Fragments
Iterator over data fragments, yielding fragment-sized data slices to be turned into payload blocks.

Type Definitions

ConfirmedFragments
Fragments for confirmed data packets.

UnconfirmedFragments
Fragments for unconfirmed data packets.