Module p25::data::header

Header generation for data packets.

A header has several user-filled fields followed by a 16-bit checksum over those fields.

Structs

BlockCount
FMF and blocks-to-follow fields.

ConfirmedFields
Header fields for confirmed packet.

ConfirmedPreamble
Preamble for confirmed data packet.

DataOffset
Byte offset into data payload where data header stops and data information begins.

Header
Builds a checksummed header based on the given fields.

HeaderPreamble
Preamble header field.

LogicalLink
Logical link ID field for addressing source or destination subscriber.

Manufacturer
Manufacturer's ID field.

PadCount
Number of pad bytes at the end of the data.

Sequencing
Syn, N(S), and FSNF fields.

ServiceAccessPoint
Service access point (SAP) field.

UnconfirmedFields
Header fields for unconfirmed packet.

UnconfirmedPreamble
Preamble for unconfirmed data packet.

Traits

BufWrite
Write some bytes into a buffer.

ByteField
Field is only a single byte.

Type Definitions

ConfirmedHeader
Packet header block for confirmed data packet.

UnconfirmedHeader
Packet header block for unconfirmed data packet.