Struct p25::data::header::Header

pub struct Header<F: BufWrite>(_);

Builds a checksummed header based on the given fields.

Methods

impl<F: BufWrite> Header<F>

fn new(fields: F) -> Header<F>

Construct a new Header with the given fields.

fn build(self) -> ([u8; 10], [u8; 2])

Get the fields and checksum that make up the header, in that order.