Struct p25::data::header::Sequencing

pub struct Sequencing {
    pub resync: bool,
    pub pkt_seq: u8,
    pub frag_seq: u8,
}

Syn, N(S), and FSNF fields.

Fields
resync: bool

Whether the receiver should resynchronize its sequence numbers using pkt_seq and frag_seq.

pkt_seq: u8

Packet sequence number, used for ordering and duplicate removal.

frag_seq: u8

Fragment sequence number, used when data is split across multiple fragments.