Function p25::data::packet::unconfirmed

pub fn unconfirmed(header: UnconfirmedHeader,
                   payload: UnconfirmedPayload)
                   -> Vec<Dibit>

Construct an unconfirmed data packet with the given header and payload blocks. The returned value is the coded, interleaved set of dibit symbols that make up the packet.