Struct p25::data::payload::Payload

pub struct Payload<'a, P: PacketParams> { /* fields omitted */ }

Wraps a buffer of bytes, splitting them over payload blocks.

Methods

impl<'a, P: PacketParams> Payload<'a, P>

fn new(data: &'a [u8]) -> Payload<'a, P>

Construct a new Payload over the given data bytes.

fn blocks(&self) -> usize

Total number of blocks in the payload.

fn pads(&self) -> usize

Number of pad bytes in the payload.

fn iter(&self) -> PayloadIter<'a, P>

Construct an iterator over the normal blocks for the payload.

fn tail(&self) -> TailBlock<'a, P>

Get the tail block of the payload.