Struct p25::data::payload::TailBlock

pub struct TailBlock<'a, P: PacketParams> { /* fields omitted */ }

Tail payload block, which has the packet checksum.

Methods

impl<'a, P: PacketParams> TailBlock<'a, P>

fn new(data: &'a [u8], checksum: u32) -> TailBlock<'a, P>

Construct a new TailBlock from the given data bytes and packet checksum.

fn build(&self) -> (&'a [u8], Range<usize>, [u8; 4])

Get the data, pad, and checksum bytes that make up the block, in that order.