Enum p25::error::P25Error

pub enum P25Error {
    RsShortUnrecoverable,
    RsMediumUnrecoverable,
    RsLongUnrecoverable,
    BchUnrecoverable,
    GolayStdUnrecoverable,
    GolayShortUnrecoverable,
    GolayExtUnrecoverable,
    HammingStdUnrecoverable,
    HammingShortUnrecoverable,
    CyclicUnrecoverable,
    ViterbiUnrecoverable,
    UnknownNid,
}

P25 runtime errors.

Variants
RsShortUnrecoverable

Too many errors were detected when attempting an RS-short decode.

RsMediumUnrecoverable

Too many errors were detected when attempting an RS-medium decode.

RsLongUnrecoverable

Too many errors were detected when attempting an RS-long decode.

BchUnrecoverable

Too many errors were detected when attempting a BCH decode.

GolayStdUnrecoverable

Too many errors were detected when attempting a standard Golay decode.

GolayShortUnrecoverable

Too many errors were detected when attempting a shortened Golay decode.

GolayExtUnrecoverable

Too many errors were detected when attempting an extended Golay decode.

HammingStdUnrecoverable

Too many errors were detected when attempting a standard Hamming decode.

HammingShortUnrecoverable

Too many errors were detected when attempting a shortened Hamming decode.

CyclicUnrecoverable

Too many errors were detected when attempting a cyclic decode.

ViterbiUnrecoverable

An ambiguous symbol or too many errors were detected when attempting convolutional decode.

UnknownNid

An unknown or corrupted NID was encountered.