Struct p25::message::data_unit::DataUnitReceiver

pub struct DataUnitReceiver { /* fields omitted */ }

State machine for low-level data unit reception.

The state machine consumes baseband samples and performs the following steps common to all data units:

Track average power of input signal

Lock onto frame synchronization

Deinterleave status symbols

Decode NID information

Decode dibit symbols until stopped

Methods

impl DataUnitReceiver

fn new() -> DataUnitReceiver

Create a new DataUnitReceiver in the initial reception state.

fn flush_pads(&mut self)

Flush any remaining padding symbols at the end of the current packet, and reenter the frame synchronization state afterwards.

fn resync(&mut self)

Force the receiver into frame synchronization.

fn feed(&mut self, s: f32) -> Option<Result<ReceiverEvent>>

Feed in a baseband symbol, possibly producing a receiver event. Return Some(Ok(event)) for any normal event, Some(Err(err)) for any error, and None if no event occurred.