Struct p25::message::nid::NIDReceiver

pub struct NIDReceiver { /* fields omitted */ }

State machine that attempts to parse a stream of dibits into an NID word.

Methods

impl NIDReceiver

fn new() -> NIDReceiver

Create a new NIDReceiver with an empty buffer.

fn feed(&mut self, dibit: Dibit) -> Option<Result<NetworkID>>

Feed in a data symbol, possibly producing a decoded NID. Return Some(Ok(nid)) if an NID was successfully parsed, Some(Err(err)) if an unrecoverable error occurred, and None for no event.