Struct p25::message::receiver::MessageReceiver

pub struct MessageReceiver {
    pub recv: DataUnitReceiver,
    // some fields omitted
}

State machine for high-level message reception.

Fields
recv: DataUnitReceiver

Lower-level stream receiver.

Methods

impl MessageReceiver

pub fn new() -> MessageReceiver

Create a new MessageReceiver in the initial state.

pub fn feed(&mut self, s: f32) -> Option<MessageEvent>

Feed in a baseband sample, possibly producing a new event or message to be handled by the given handler.