Trait p25::message::receiver::MessageHandler

pub trait MessageHandler {
    fn handle_error(&mut self, recv: &mut DataUnitReceiver, err: P25Error);
    fn handle_nid(&mut self, recv: &mut DataUnitReceiver, nid: NetworkID);
    fn handle_header(&mut self, recv: &mut DataUnitReceiver, header: VoiceHeaderFields);
    fn handle_frame(&mut self, recv: &mut DataUnitReceiver, frame: VoiceFrame);
    fn handle_lc(&mut self, recv: &mut DataUnitReceiver, lc: LinkControlFields);
    fn handle_cc(&mut self, recv: &mut DataUnitReceiver, cc: CryptoControlFields);
    fn handle_data_frag(&mut self, recv: &mut DataUnitReceiver, data: u32);
    fn handle_tsbk(&mut self, recv: &mut DataUnitReceiver, tsbk: TSBKFields);
    fn handle_term(&mut self, recv: &mut DataUnitReceiver, lc: Option<LinkControlFields>);
}

Set of callbacks to handle events and messages that occur when receiving the air interface.

Required Methods

fn handle_error(&mut self, recv: &mut DataUnitReceiver, err: P25Error)

A runtime error occured.

fn handle_nid(&mut self, recv: &mut DataUnitReceiver, nid: NetworkID)

An NID was decoded.

fn handle_header(&mut self, recv: &mut DataUnitReceiver, header: VoiceHeaderFields)

A voice header was received.

fn handle_frame(&mut self, recv: &mut DataUnitReceiver, frame: VoiceFrame)

A voice frame was received.

fn handle_lc(&mut self, recv: &mut DataUnitReceiver, lc: LinkControlFields)

A link control word was decoded.

fn handle_cc(&mut self, recv: &mut DataUnitReceiver, cc: CryptoControlFields)

A crypto control word was decoded.

fn handle_data_frag(&mut self, recv: &mut DataUnitReceiver, data: u32)

A voice low-speed data fragment was decoded.

fn handle_tsbk(&mut self, recv: &mut DataUnitReceiver, tsbk: TSBKFields)

A trunking signalling packet was received.

fn handle_term(&mut self, recv: &mut DataUnitReceiver, lc: Option<LinkControlFields>)

A voice terminator packet was received, optionally with link control word.