Module p25::trunking::fields

Decoders for various trunking-related packet fields.

Structs

AdjacentSite
Advertisement of an adjacent/nearby site within the same WACN (Wide Area Communication Network.)

Channel
Uniquely identifies a channel within a site.

ChannelParams
Computes TX/RX frequencies and bandwidth for channel numbers within a site.

ChannelParamsUpdate
Advertisement of parameters used to calculate TX/RX frequencies within the given associated channel.

ServiceOptions

SiteOptions
Options for a P25 site.

SystemServices

Enums

TalkGroup
Identifies which group a message belongs to.

Functions

parse_updates
Parse out the pair of channels/talkgroups found in a Group Voice Update.

Type Definitions

ChannelParamsMap
Map channel identifier (maximum 16 per control channel) to its parameters.

ChannelUpdates
Represents the channel updates seen in a Group Voice Update.