Struct p25::trunking::fields::AdjacentSite

pub struct AdjacentSite<'a>(_);

Advertisement of an adjacent/nearby site within the same WACN (Wide Area Communication Network.)

Methods

impl<'a> AdjacentSite<'a>

fn new(payload: &'a [u8]) -> Self

Create a new AdjacentSite decoder from the given payload bytes.

fn area(&self) -> u8

Location registration area of adjacent site, which determines whether a subscriber must update the network before roaming to the site.

fn opts(&self) -> SiteOptions

Description of adjacent site.

fn system(&self) -> u16

System ID of adjacent site within WACN.

fn rfss(&self) -> u8

RF Subsystem ID of adjacent site within the System.

fn site(&self) -> u8

Site ID of adjacent site within the RFSS.

fn channel(&self) -> Channel

Channel information for computing TX/RX frequencies.

fn services(&self) -> SystemServices

Services supported by the adjacent site.