Struct p25::trunking::fields::GroupTrafficUpdate

pub struct GroupTrafficUpdate<'a>(_);

Updates subscribers about new or ongoing talkgroup conversations.

Note that this can be used for both GroupVoiceUpdate and GroupDataUpdate.

Methods

impl<'a> GroupTrafficUpdate<'a>

pub fn new(payload: &'a [u8]) -> Self

Create a new GroupTrafficUpdate decoder from the given payload bytes.

pub fn updates(&self) -> [(Channel, TalkGroup); 2]

Retrieve the set of active talkgroups included in the update along with the parameters for tuning to the traffic channel of each.