Module p25::trunking::tsbk

Receive Trunking Signalling Block (TSBK) packets and decode the various TSBK payloads.

Structs

GroupDataGrant
Indicates a talkgroup has been granted a data traffic channel.

GroupVoiceGrant
Indicates a talkgroup has been granted a voice traffic channel.

LocRegResponse
Response given to a location registration request.

PhoneGrant
Indicates a unit has been granted a traffic channel for a phone call.

TSBKFields
A Trunking Signalling Block packet.

TSBKReceiver
State machine for receiving a TSBK packet.

UnitDeregAck
Acknowledgement of successful user deregistration request.

UnitRegResponse
Response given to an attempted user registration.

UnitTrafficChannel
Indicates a pair of units have been granted a traffic channel.

Enums

TSBKOpcode
Type of a TSBK payload.

Type Definitions

Buf
Buffer of bytes that represents a TSBK packet.