Struct p25::trunking::tsbk::PhoneGrant

pub struct PhoneGrant(_);

Indicates a unit has been granted a traffic channel for a phone call.

Methods

impl PhoneGrant

fn new(tsbk: TsbkFields) -> Self

Create a new PhoneGrant decoder from the base TSBK decoder.

fn opts(&self) -> ServiceOptions

Options requested/granted for the traffic channel.

fn channel(&self) -> Channel

Parameters for tuning to the traffic channel.

fn call_timer(&self) -> u16

Maximum amount of time (in units of 100ms) that the phone call can occupy the traffic channel.

fn unit(&self) -> u32

Unit assigned to the call.