Struct p25::trunking::tsbk::TSBKFields

pub struct TSBKFields(_);

A Trunking Signalling Block packet.

Methods

impl TSBKFields

fn new(buf: Buf) -> TSBKFields

Interpret the given bytes as a TSBK packet.

fn is_tail(&self) -> bool

Whether this packet is the last one in the TSBK group.

fn protected(&self) -> bool

Whether the packet is encrypted.

fn opcode(&self) -> Option<TSBKOpcode>

Type of data contained in the payload.

fn mfg(&self) -> u8

Manufacturer ID, which determines if the packet is standardized.

fn crc(&self) -> u16

Transmitted CRC.

fn calc_crc(&self) -> u16

Calculate 16-bit CRC over bytes in packet.

fn crc_valid(&self) -> bool

Verify if the calculated CRC matches the transmitted one.

fn payload(&self) -> &[u8]

Bytes that make up the payload of the packet.