Struct p25::trunking::tsbk::UnitRegResponse

pub struct UnitRegResponse(_);

Response given to an attempted user registration.

Methods

impl UnitRegResponse

fn new(tsbk: TsbkFields) -> Self

Create a new UnitRegResponse decoder from the base TSBK decoder.

fn response(&self) -> RegResponse

System response to user registration request.

fn system(&self) -> u16

System ID within WACN.

fn src_id(&self) -> u32

Address of originating unit, which uniquely identifies the unit within the System.

fn src_addr(&self) -> u32

ID of originating unit which, along with the WACN and System ID, uniquely identifies the unit.