Module p25::voice::frame_group

Receive voice frame groups, known as LDU1 and LDU2 in the standard.

Each frame group contains 9 voice frames, a low-speed data word, and an "extra" packet: either link control (LC) or crypto control (CC).

Structs

CryptoControlExtra
Crypto control frame group extra.

FrameGroupReceiver
State machine that receives the various pieces that make up a frame group.

LinkControlExtra
Link control frame group extra.

Enums

FrameGroupEvent
Events that can occur when receiving a frame group.

Traits

Extra
An "extra" information packet carried along in a frame group.

Type Definitions

VoiceCCFrameGroupReceiver
Receiver for Crypto Control (CC) frame group.

VoiceLCFrameGroupReceiver
Receiver for Link Control (LC) frame group.