Module p25::data

Implements Project 25's data packet specification.

Reexports

pub use self::fragment::{ConfirmedFragments, UnconfirmedFragments};

pub use self::header::{ConfirmedHeader, ConfirmedFields, ConfirmedPreamble, UnconfirmedHeader, UnconfirmedFields, UnconfirmedPreamble, ServiceAccessPoint, Manufacturer, LogicalLink, BlockCount, PadCount, Sequencing, DataOffset};

pub use self::payload::{ConfirmedPayload, UnconfirmedPayload};

Modules

coder
Provides a convenience interface for coding symbols into a buffer.

crc
Implements the 9, 16, and 32-bit CRCs defined by P25 for data checksums.

fields
Data packet fields.

fragment
Provides Fragments iterator for splitting a data slice into fragment slices, whose size depends on the data packet used.

header
Header generation for data packets.

interleave
Interleaving and deinterleaving for data packet payloads.

packet
Generate data packets.

params
Defines various parameters used for data packets.

payload
Generate data blocks from a slice of bytes.