Enum p25::error::P25Error

pub enum P25Error {
    ReedSolomonUnrecoverable,
    BCHUnrecoverable,
    GolayUnrecoverable,
    HammingUnrecoverable,
    CyclicUnrecoverable,
    ViterbiUnrecoverable,
    UnknownNID,
}

P25 runtime errors.

Variants
ReedSolomonUnrecoverable

Too many errors were detected when attempting a RS decode.

BCHUnrecoverable

Too many errors were detected when attempting a BCH decode.

GolayUnrecoverable

Too many errors were detected when attempting a Golay decode.

HammingUnrecoverable

Too many errors were detected when attempting a Hamming decode.

CyclicUnrecoverable

Too many errors were detected when attempting a cyclic decode.

ViterbiUnrecoverable

An ambiguous symbol or too many errors were detected when attempting convolutional decode.

UnknownNID

An unknown or corrupted NID was encountered.