Module p25::trunking::fields

Decode various trunking-related packet fields.

Structs

AdjacentSite
Advertisement of an adjacent/nearby site within the same WACN (Wide Area Communication Network.)

AltControlChannel
Advertisement of one or more alternative control channels for the current site.

Channel
Uniquely identifies a channel within a site.

ChannelParams
Computes TX/RX frequencies and bandwidth for channel numbers within a site.

ChannelParamsUpdate
Advertisement of parameters used to calculate TX/RX frequencies within the given associated channel.

GroupTrafficUpdate
Updates subscribers about new or ongoing talkgroup conversations.

NetworkStatusBroadcast
WACN (Wide Area Communication Network) and System ID information of current control channel.

PhoneAlert
Alerts a unit of a call from the public phone network.

RFSSStatusBroadcast
Site and RFSS information of current control channel.

ServiceOptions
Options that can be requested/granted by a service.

SiteOptions
Options for a P25 site.

SystemServices
Supported services of a control channel.

UnitCallAlert
Request for a target unit to call a source unit.

UnitCallRequest
Signals a target unit that a unit-to-unit all has been requested.

Enums

RegResponse
Registration response.

TalkGroup
Identifies which group a message belongs to.

Type Definitions

ChannelParamsMap
Maps channel identifiers (maximum 16 per control channel) to their tuning parameters.