Struct p25::voice::frame_group::FrameGroupReceiver

pub struct FrameGroupReceiver<E: Extra> { /* fields omitted */ }

State machine that receives the various pieces that make up a frame group.

Methods

impl<E: Extra> FrameGroupReceiver<E>

fn new() -> FrameGroupReceiver<E>

Create a new FrameGroupReceiver in the initial state.

fn done(&self) -> bool

Whether the full frame group has been received.

fn feed(&mut self, dibit: Dibit) -> Option<Result<FrameGroupEvent<E>>>

Feed in a baseband symbol, possibly producing an event. Return Some(Ok(event)) if a nominal event occurred, Some(Err(err)) if an error occurred, and None in the case of no event.